Changed logic
switch case is now outside the while loop when connecting to the database. This increases the amount of while loops, but ensures that the correct query is selected. However, there still seems to be a problem that does not allow for any case other than default to be executed. This needs to be reviewed and fixed. 

Also added placeholders for item select and item modify statements, along with functions that will enable the functioning of these statements.